Complete redesign of the Windows cross compilation script
The script now also supports changing the compilation target. The
recommended targets are i686-w64-mingw32.static for 32bit binaries and
x86_64-w64-mingw32.static for 64bit binaries.

The script is redesigned with functions to be called from within the build
system script. To invoke it on it's own you can specify the --compile and
--compile-dependencies arguments.