The build system now asks which tag version you want to build
Implemented by adding a choose_tags() function that brings up a radiolist
dialog with all the detected tags in the repository.

The function checkout() was also added. This function simply  checks out
the selected version (so it requires that choose_tags() has been run in
advance). If the checkout fails, it informs the user of this fact.