Add "macdeployqtplus" to the build system
This tiny application rewrites a compiled OSX application and prepares
it for packaging in a DMG - rewriting a dynamically compiled program and
its library paths to work in a distributed DMG package.

This was taken from the contrib/ directory in Bitcoin and has also been
slightly improved compared to that version.